Restore previous convert visitor approach to correctly bind columns
Instead of attempting to get away without using reflection (and then resorting to it anyways),
this setup reuses the previously used approach to updating the SelectExpression.
Due to internal changes that means a slightly different invocation of the GenerateOuterColumnAccessor
preceded by a reflection-based restoration of the _sqlAliasManager that is needed for the call of
PushDownIntoSubQuery. Future versions of .NET may make this obsolete by removing the use of
SqlAliasManager within SelectExpression entirely, but that's a future concern.